+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - none
 
 # v2026.8.1-beta
-- sqlmath - Update c-function winSinefitSnr() with error-handling for singularity.
+- sqlmath - Remove obsolete sql-functions MEDIAN2(), QUANTILE(), superseded by sqlite-builtins.
+- sqlmath - Update c-function winSinefitSnr() with error-handling for NaN.
 - sqlmath - Fix c-header conflict between dlfcn.h and Python.h.
 
 # v2026.7.30